add initial data

+++ src/main/java/ch/zli/m223/model/BookingTyp.java
@@ -6,5 +6,5 @@ public interface BookingTyp {
 
     String getBookingType();
 
-    Float getPrice();
+    Double getPrice();
 }
