+++ src/main/java/ch/zli/m223/model/Subscription.java
@@ -5,5 +5,5 @@ public interface Subscription {
 
     String getSubscriptionType();
 
-    Integer getPrice();
+    Float getPrice();
 }
